+++ files/layout.blade.php
@@ -87,7 +87,7 @@
                 <li>
                   <a href="{{ url('logout') }}">
                     <i class="fa fa-btn fa-sign-out"></i>
-                    Logout
+                    ログアウト
                   </a>
                 </li>
               </ul>